python-llm-service/requirements.txt
# Web Framework
Flask==2.2.5
Werkzeug==2.2.3

# Core Dependencies
requests==2.28.2
python-dotenv==0.21.1

# AI and Language Models
openai==0.27.0
pydantic==1.10.12
langchain==0.0.118